More Optimizations
Make some further runtime optimizations based on Python's quirks as a
language. As a rule, we want to avoid incrementing variables, as
python treats variables as objects, so incrementing is
extremely expensive. Try to use for range loops as much as possible.

Runtime in testing still comes out to almost double TimSort's time
when half of our data is duplicated, which isn't bad... but it's not
perfect.